sirio: qual
Unify the 'len' variable: there are numerous places in the code where
we depend on having used a certain logic to calculate the length of
'buf' - so don't court disaster by setting this in multiple places.

Also, change the error checking so that the print statement on error
shows that 'malloc' was being executed.